+++ frontend/components/custom-table/department/columns.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import { columns as courseDegreeColumns } from "../main/columns";
+
+const toOmit = ["department.name", "university.name", "university.city"];
+
+export const columns = courseDegreeColumns.filter(
+  // @ts-ignore this property normally is not accessed 
+  (column) => !toOmit.includes(column.accessorKey ?? ""),
+);
